+++ src/stores/UserStore.js
@@ -1,7 +1,12 @@
 import { defineStore } from 'pinia';
 
 export const useUserStore = defineStore('user', {
-  state: () => ({ username: '', firstname: '' }),
+  state: () => ({
+    id: '',
+    username: '',
+    firstname: '',
+    roles: [],
+  }),
   getters: {
     isEmpty: (state) => state.username.length === 0 && state.firstname.length === 0,
   },